chore: design doc updates and types and constants
